Fixed catalog rule group website query on PostgreSQL
Applying rules got past the price chain but then died in
_prepareGroupWebsite() with "CASE types boolean and integer cannot be
matched": the open-ended to_time window was expressed as a CASE
returning either 1 or a comparison. It is a plain disjunction.

Also replaced StrictGroupByTest, which reflected into the removed
_prepareIndexSelect(), with coverage of the aggregation _fillIndexData()
now performs, running on every backend instead of MySQL only. Moved
PriceChainTest next to it under Integration/ since it drives the whole
indexer against the live database, typed _fillIndexData(), and routed
the indexed dates through Locale::formatDateForDb().
