+++ .phpstan.dist.baseline.neon
@@ -11182,12 +11182,6 @@ parameters:
 			path: app/code/core/Mage/CatalogRule/Model/Action/Index/Refresh.php
 
 		-
-			rawMessage: 'Method Mage_CatalogRule_Model_Action_Index_Refresh::_fillIndexData() has no return type specified.'
-			identifier: missingType.return
-			count: 1
-			path: app/code/core/Mage/CatalogRule/Model/Action/Index/Refresh.php
-
-		-
 			rawMessage: 'Method Mage_CatalogRule_Model_Action_Index_Refresh::_getProduct() has no return type specified.'
 			identifier: missingType.return
 			count: 1
